requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.11.4
qrcode==8.1
python-jose==3.3.0
passlib==1.7.4
pymongo==4.6.0
python-dotenv==1.0.0
email-validator==2.2.0
python-multipart==0.0.6
bcrypt==4.0.1
pillow==10.2.0  # For QR code image handling